refactored to getGGW
